prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,9 +10,9 @@
 
 void fibonacci(void)
 {
-	long long int i = 1;
-	long long  int j = 2;
-	long long int tmp;
+	long int i = 1;
+	long  int j = 2;
+	long int tmp;
 	int counter;
 
 	for (counter = 0; counter < 50; counter++)
